+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 12.10.2015  
+OTER:
+
+- Add Utils namespace for utilite methods. Add ToLower method.
+- Add Settings class.
+- Work on UML class diagram. 
 ### 11.10.2015  
 OTER: